+++ src/ui/mod.rs
@@ -1,3 +1,4 @@
+pub mod about;
 pub mod config_window;
 pub mod password_field;
 pub mod popup;